# perf: pregenerate `timeWindow` string when possible