We shouldn't keep recreating the string if it's static, it might become a problem under an attack

**Note:** This was already the case, but it got removed in https://github.com/fastify/fastify-rate-limit/pull/357