@gurgunday 2024-02-24
Uzlopak was right, noop is faster than ?.